Double изменен на Decimal при вычислениях

+++ MathParserWPF/Model/MathInterpreter.cs
@@ -27,19 +27,19 @@ namespace MathParserWPF.Model
         }
 
         // рекурсивный метод, который вызывается для каждого узла дерева
-        private double ExecuteNode(AstNode node)
+        private decimal ExecuteNode(AstNode node)
         {
             if (node.NodeType == AstNode.Type.Unknown)
                 throw new InterpreterException("Неопределенный тип узла AST-дерева");
             if (node.NodeType == AstNode.Type.Number)
             {
                 if (node.Text.Length > 12) throw new ComputingException("Программа не может обработать такое большое или маленькое число.");
-                bool ok = double.TryParse(node.Text, NumberStyles.Any, NFI, out double res);
+                bool ok = decimal.TryParse(node.Text, NumberStyles.Any, NFI, out decimal res);
                 if (!ok) throw new ComputingException("Программа не может обработать такое большое или маленькое число.");
                 return res;
             }
-            double a = ExecuteNode(node.GetChild(0));
-            double b = ExecuteNode(node.GetChild(1));
+            decimal a = ExecuteNode(node.GetChild(0));
+            decimal b = ExecuteNode(node.GetChild(1));
             switch (node.NodeType)
             {
                 case AstNode.Type.Add:
@@ -56,12 +56,12 @@ namespace MathParserWPF.Model
             }
         }
         // public-метод для вызова интерпретации
-        public double Execute()
+        public decimal Execute()
         {
             return ExecuteNode(_programNode);
         }
         // статическая реализации предыдузего метода (для удобства)
-        public static double Execute(AstNode programNode)
+        public static decimal Execute(AstNode programNode)
         {
             MathInterpreter mei = new MathInterpreter(programNode);
             return mei.Execute();
